# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rproxy-api control API",
    "version": "see GET /capabilities and the release",
    "description": "The contract is docs/API.md; this document follows it."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Protocol": {
        "type": "string",
        "enum": [
          "tcp",
          "udp"
        ],
        "description": "Case-insensitive on input."
      },
      "SourceIp": {
        "type": "string",
        "enum": [
          "proxy",
          "proxy_v1",
          "proxy_v2",
          "transparent"
        ]
      },
      "StartTls": {
        "type": "string",
        "enum": [
          "smtp",
          "imap",
          "pop3"
        ]
      },
...
      "Capabilities": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "The rproxy-api release of this build (Cargo.toml version, e.g. 0.3.18). Added in 0.3.18; absent on older releases",
            "example": "0.3.18"
          },
          "source_ip": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceIp"
            }
          },
          "transparent": {
            "type": "boolean"
          },
          "transparent_ipv6": {
            "type": "boolean"
          },
          "tls_modes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dtls": {
            "type": "boolean"
          },
          "starttls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StartTls"
            }
          },
          "max_range_ports": {
            "type": "integer"
          },
          "features": {
            "type": "object",
            "properties": {
              "http": {
                "type": "boolean"
              },
              "http3": {
                "type": "boolean"
              },
              "acme": {
                "type": "boolean"
              },
              "tls_options": {
                "type": "boolean"
              },
              "middlewares": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "services": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/capabilities": {
      "get": {
        "summary": "What this build and host can do (any valid token)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "This document (any valid token)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/interfaces": {
      "get": {
        "summary": "Addresses rules can listen on (rules:read)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interfaces"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/config": {
      "get": {
        "summary": "State of the settings file RPROXY_CONFIG (rules:read)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/config/reload": {
      "post": {
        "summary": "Apply the settings file RPROXY_CONFIG now and answer what happened (admin; by default only over the Unix socket, RPROXY_API_RELOAD_UNIX_ONLY)",
        "responses": {
          "200": {
            "description": "Applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReload"
                }
              }
            }
          },
          "400": {
            "description": "The settings have a mistake; nothing changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReloadError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the admin scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "No settings file is configured (code no_config)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/rules": {
      "get": {
        "summary": "All rules (rules:read)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a rule (rules:write; with an acme certificate also acme:write)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "400": {
            "description": "invalid / tls_config / unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "already_exists / reserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "resolve_failed / bind_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/rules/{protocol}/{listen_addr}/{listen_port}": {
      "parameters": [
        {
          "name": "protocol",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Protocol"
          }
        },
        {
          "name": "listen_addr",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "IP address; IPv6 without brackets"
        },
        {
          "name": "listen_port",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 65535
          }
        }
      ],
      "get": {
        "summary": "One rule (rules:read)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Change a rule; new connections use it at once (rules:write; with an acme certificate also acme:write)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "400": {
            "description": "invalid / tls_config / unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "static",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a rule (rules:write)",
        "parameters": [
          {
            "name": "drain_secs",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Wait this long for connections to finish before closing them"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "static",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks a scope or the listen port",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/acme": {
      "get": {
        "summary": "ACME accounts, DNS providers, resolvers and certificates (rules:read); names and states only, never a secret",
        "responses": {
          "200": {
            "description": "global.acme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acme"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/acme/renew": {
      "post": {
        "summary": "Renew an ACME certificate now, within the rate limit (acme:write; by default only over the Unix socket)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolver": {
                    "type": "string"
                  },
                  "domains": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "resolver",
                  "domains"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The renewal is started in the background (see GET /acme)"
          },
          "400": {
            "description": "Unknown resolver, or names outside the allowlists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the acme:write scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No rule uses this certificate, or global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/acme/accounts/{name}/register": {
      "post": {
        "summary": "Create the account at the CA, or find the existing one for its key (acme:write; by default only over the Unix socket)",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "An account of global.acme.accounts"
          }
        ],
        "responses": {
          "200": {
            "description": "Done"
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the acme:write scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such account, or global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The CA refused or could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}